git_update.py: Do not print to stderr on success
We use a script that sends emails for cron job runs if they exited with
a non-zero error code or printed something on stderr. Unfortunately,
git's progress information is printed to stderr, which means we
currently get an email for every update run.

Silence this by passing --quiet to git in a couple of places.